Introduce GhostBoundaryElements
Resolve an edge case where a parent face coincident to an internal boundary, with only ghost children
would not have essential boundary conditions applied on DOF that were not members of the ghost faces.
The effect of this was to not constrain face DOFs for higher order finite elements.

This change was achieved by introducing a map from boundary attribute to sets of parent faces that
are secretly members of the boundary element set. These are termed ghost boundary elements and are computed
during ExchangeFaceNbrData, and then used by the ParFiniteElementSpace if there is a non conformal mesh.